qwen_infer/requirements.txt
accelerate==1.11.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
Bottleneck==1.4.2
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
datasets==3.3.2
diffusers==0.36.0.dev0
dill==0.3.8
exceptiongroup==1.3.0
filelock==3.17.0
frozenlist==1.8.0
fsspec==2024.12.0
gmpy2==2.2.1
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.34.4
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mkl_fft==2.0.0
mkl_random==1.2.11
mkl-service==2.5.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.4.2
numexpr==2.14.1
numpy==2.2.5
optree==0.17.0
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
pip==25.2
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==21.0.0
pybind11==3.0.1
pybind11-global==3.0.1
pycares==4.10.0
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
torchvision-extra-decoders==0.0.2
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0+git576374f8
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvloop==0.22.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.22.0
zipp==3.23.0
